Fixed problems associated with subclassed User and Group collections and auth functions and roles
