+++ pfunk/__init__.py
@@ -5,8 +5,9 @@
 .. include:: ../CONTRIBUTE.md
 """
 __docformat__ = "google"
+
+from .client import FaunaClient
 from .collection import Collection, Enum
 from .fields import (StringField, IntegerField, DateField, DateTimeField, BooleanField, FloatField, EmailField,
                      EnumField, ReferenceField, ManyToManyField, SlugField)
 from .project import Project
-from .client import FaunaClient
