+++ pfunk/contrib/auth/collections/user.py
@@ -1,101 +1,21 @@
-import datetime
-import json
-import random
 import uuid
 
-import jwt
-from cryptography.fernet import Fernet
-from dateutil import tz
 from envs import env
-from faunadb.errors import BadRequest, NotFound
-from jwt import ExpiredSignatureError
+from faunadb.errors import BadRequest
 from valley.exceptions import ValidationException
 from valley.utils import import_util
-from werkzeug.utils import cached_property
 
 from pfunk.client import q
 from pfunk.collection import Collection, Enum
+from pfunk.contrib.auth.collections import Key
 from pfunk.contrib.auth.resources import LoginUser, UpdatePassword, Public, UserRole, LogoutUser
-from pfunk.contrib.auth.views import ForgotPasswordChangeView, LoginView, SignUpView, VerifyEmailView, LogoutView, UpdatePasswordView, ForgotPasswordView
+from pfunk.contrib.auth.views import ForgotPasswordChangeView, LoginView, SignUpView, VerifyEmailView, LogoutView, \
+    UpdatePasswordView, ForgotPasswordView
 from pfunk.contrib.email.base import send_email
-from pfunk.exceptions import LoginFailed, DocNotFound, Unauthorized
-from pfunk.fields import EmailField, SlugField, ManyToManyField, ListField, ReferenceField, StringField, EnumField
+from pfunk.exceptions import LoginFailed, DocNotFound
+from pfunk.fields import EmailField, ManyToManyField, StringField, EnumField
 
 AccountStatus = Enum(name='AccountStatus', choices=['ACTIVE', 'INACTIVE'])
-
-
-class Key(object):
-
-    @classmethod
-    def create_keys(cls):
-        c = cls()
-        keys = {}
-        for i in range(10):
-            kid = str(uuid.uuid4())
-            k = {'signature_key': Fernet.generate_key().decode(), 'payload_key': Fernet.generate_key().decode(),
-                 'kid': kid}
-            keys[kid] = k
-        return keys
-
-    @classmethod
-    def import_keys(cls):
-        try:
-            keys = import_util(env('KEY_MODULE', 'bad.import'))
-        except ImportError:
-            keys = {}
-        return keys
-
-    @classmethod
-    def get_keys(cls):
-        keys = cls.import_keys()
-        return list(keys.values())
-
-    @classmethod
-    def get_key(cls):
-
-        return random.choice(cls.get_keys())
-
-    @classmethod
-    def create_jwt(cls, secret_claims):
-
-        key = cls.get_key()
-        pay_f = Fernet(key.get('payload_key'))
-        gmt = tz.gettz('GMT')
-        now = datetime.datetime.now(tz=gmt)
-        exp = now + datetime.timedelta(days=1)
-        payload = {
-            'iat': now.timestamp(),
-            'exp': exp.timestamp(),
-            'nbf': now.timestamp(),
-            'iss': env('PROJECT_NAME', 'pfunk'),
-            'til': pay_f.encrypt(json.dumps(secret_claims).encode()).decode()
-        }
-        return jwt.encode(payload, key.get('signature_key'), algorithm="HS256", headers={'kid': key.get('kid')}), exp
-
-    @classmethod
-    def decrypt_jwt(cls, encoded):
-        headers = jwt.get_unverified_header(encoded)
-        keys = cls.import_keys()
-        key = keys.get(headers.get('kid'))
-        try:
-            decoded = jwt.decode(encoded, key.get('signature_key'), algorithms="HS256", verify=True,
-                             options={"require": ["iat", "exp", "nbf", 'iss', 'til']})
-        except ExpiredSignatureError:
-            raise Unauthorized('Unauthorized')
-        pay_f = Fernet(key.get('payload_key').encode())
-        k = pay_f.decrypt(decoded.get('til').encode())
-        return json.loads(k.decode())
-
-
-class Group(Collection):
-    """ Group collection that the user belongs to """
-    name = StringField(required=True)
-    slug = SlugField(unique=True, required=False)
-    users = ManyToManyField(
-        'pfunk.contrib.auth.collections.User', relation_name='users_groups')
-
-    def __unicode__(self):
-        return self.name  # pragma: no cover
 
 
 def attach_verification_key(doc):
@@ -119,9 +39,10 @@ class BaseUser(Collection):
     collection_roles = [Public, UserRole]
     non_public_fields = ['groups']
     use_email_verification = True
-    group_class = env('GROUP_COLLECTION', 'pfunk.contrib.auth.collections.Group')
+    group_class = import_util(env('GROUP_COLLECTION', 'pfunk.contrib.auth.collections.group.Group'))
     # Views
-    collection_views = [LoginView, SignUpView, VerifyEmailView, LogoutView, UpdatePasswordView, ForgotPasswordView, ForgotPasswordChangeView]
+    collection_views = [LoginView, SignUpView, VerifyEmailView, LogoutView, UpdatePasswordView, ForgotPasswordView,
+                        ForgotPasswordChangeView]
     # Signals
     pre_create_signals = [attach_verification_key]
     post_create_signals = [send_verification_email]
@@ -148,7 +69,7 @@ class BaseUser(Collection):
         try:
             return c.client(_token=_token).query(
                 q.call("login_user", {
-                       "username": username, "password": password})
+                    "username": username, "password": password})
             )
         except BadRequest:
             raise LoginFailed(
@@ -161,6 +82,7 @@ class BaseUser(Collection):
         return c.client(_token=_token).query(
             q.call("logout_user")
         )
+
 
     def permissions(self, _token=None):
         return []
@@ -245,7 +167,7 @@ class BaseUser(Collection):
 
     @classmethod
     def forgot_password(cls, email):
-        """ Sends forgot password email to let user 
+        """ Sends forgot password email to let user
             use that link to reset their password
         """
         user = cls.get_by('unique_User_email', email)
@@ -321,65 +243,11 @@ class BaseUser(Collection):
         return self.username  # pragma: no cover
 
 
-class UserGroups(Collection):
-    """ Many-to-many collection of the user-group relationship
-
-        The native fauna-way of holding many-to-many relationship
-        is to only have the ID of the 2 object. Here in pfunk, we
-        leverage the flexibility of the collection to have another
-        field, which is `permissions`, this field holds the capablities
-        of a user, allowing us to add easier permission handling.
-        Instead of manually going to roles and adding individual
-        collections which can be painful in long term.
-
-    Attributes:
-        collection_name (str):
-            Name of the collection in Fauna
-        userID (str):
-            Fauna ref of user that is tied to the group
-        groupID (str):
-            Fauna ref of a collection that is tied with the user
-        permissions (str[]):
-            List of permissions, `['create', 'read', 'delete', 'write']`
-    """
-    collection_name = 'users_groups'
-    userID = ReferenceField(env('USER_COLLECTION', 'pfunk.contrib.auth.collections.User'))
-    groupID = ReferenceField(env('GROUP_COLLECTION', 'pfunk.contrib.auth.collections.Group'))
-    permissions = ListField()
-
-    def __unicode__(self):
-        return f"{self.userID}, {self.groupID}, {self.permissions}"
-
-
-class PermissionGroup(object):
-    """ List of permission that a user/object has
-
-    Attributes:
-        collection (`pfunk.collection.Collection`, required):
-            Collection to allow permissions
-        permission (list, required):
-            What operations should be allowed `['create', 'read', 'delete', 'write']`
-    """
-    valid_actions: list = ['create', 'read', 'delete', 'write']
-
-    def __init__(self, collection: Collection, permissions: list):
-        if not issubclass(collection, Collection):
-            raise ValueError(
-                'Permission class requires a Collection class as the first argument.')
-        self.collection = collection
-        self._permissions = permissions
-        self.collection_name = self.collection.get_class_name()
-
-    @cached_property
-    def permissions(self):
-        """ Lists all collections and its given permissions """
-        return [f'{self.collection_name}-{i}'.lower() for i in self._permissions if i in self.valid_actions]
-
-
 class User(BaseUser):
+    user_group_class = import_util('pfunk.contrib.auth.collections.common.UserGroups')
+    group_class = import_util('pfunk.contrib.auth.collections.group.Group')
     """ User that has permission capabilities. Extension of `BaseUser` """
-    groups = ManyToManyField(env('GROUP_COLLECTION', 'pfunk.contrib.auth.collections.Group'), 'users_groups')
-
+    groups = ManyToManyField(env('GROUP_COLLECTION', 'pfunk.contrib.auth.collections.group.Group'), 'users_groups')
 
     @classmethod
     def get_permissions(cls, ref, _token=None):
@@ -407,8 +275,8 @@ class User(BaseUser):
         """
         perm_list = []
         for i in self.get_groups(_token=_token):
-            ug = UserGroups.get_index('users_groups_by_group_and_user', [
-                                      i.ref, self.ref], _token=_token)
+            ug = self.user_group_class.get_index('users_groups_by_group_and_user', [
+                i.ref, self.ref], _token=_token)
             for user_group in ug:
                 p = []
                 if isinstance(user_group.permissions, list):
@@ -418,24 +286,24 @@ class User(BaseUser):
         return perm_list
 
     def add_permissions(self, group, permissions: list, _token=None):
-        """ Adds permission for the user 
-        
-        Adds permission by extending the list of permission 
-        in the many-to-many collection of the user, i.e. in 
+        """ Adds permission for the user
+
+        Adds permission by extending the list of permission
+        in the many-to-many collection of the user, i.e. in
         the `UserGroup` collection.
 
         Args:
-            group (str, required): 
+            group (str, required):
                 Group collection of the User
             permissions (list, required):
                 Permissions to give, `['create', 'read', 'delete', 'write']`
                 Just add the operation you need
             _token (str, required):
                 auth token of the user
-        
+
         Returns:
             UserGroup (`contrib.auth.collections.UserGroup`):
-                `UserGroup` instance which has the added permissions 
+                `UserGroup` instance which has the added permissions
                 of the user
         """
         perm_list = []
@@ -443,9 +311,9 @@ class User(BaseUser):
             perm_list.extend(i.permissions)
 
         try:
-            user_group = UserGroups.get_by('users_groups_by_group_and_user', terms=[group.ref, self.ref])
+            user_group = self.user_group_class.get_by('users_groups_by_group_and_user', terms=[group.ref, self.ref])
         except DocNotFound:
-            user_group = UserGroups.create(userID=self.ref, groupID=group.ref, permissions=perm_list)
+            user_group = self.user_group_class.create(userID=self.ref, groupID=group.ref, permissions=perm_list)
         if user_group.permissions != perm_list:
             user_group.permissions = perm_list
         user_group.save()
