+++ pfunk/queryset.py
